Updated HDBDOS makefile so that the assembler is now a variable rather than a fixed executable.
Edited HDBDOS makefile so that a variable called "AS" is now set
to the assembler to be used.  Then all sections were updated so
they use $(AS) inplace of "mamou".

This should allow for easier change out of the assembler in the
future as well as any special assembler options.