Updated cecb copy -d -e options. Updated hdbdos makefile to create two cas files.
Updated cecb copy command line options -d & -e.  These would normally
cause a unknown - command when used.  This was because strtol was used
and would set the option pointer to the null char.  This was causing
problem with command line scan routine as it looked for a null char
using a for().  Since the for() would increase the char pointer past
the null char it would error out.  To correct this before doing the
"break;" a "p=p-1;" was added right before the "break;" to help
resolve this problem.

Updated the hdbdos makefile to create two cas files for use with
emulators.  These two files are hdbdw3bck.cas and hdbdw3bc3.cas