dwdos: makewav can read start/exec addresses from decb binary
I also believe the generated WAV's used to be broken because the
binaries were compiled to have DECB header/footer and the header was
previously baked into the WAV file as normal data. The exec address
would then be wrong.

The new makewav (with -c option) will parse the DECB header/footer and
not include them in the data.