hdbdos: makefile: Factor out mamou-specific options
So that they can be overridden. For example:

 $ make AS="lwasm -r --pragma=condundefzero,nodollarlocal" \
        CECB=../build/unix/cecb/cecb

Also make use of the new -c option of makewav so that it
parses the DECB headers and sets start/exec addresses.

At this point lwasm and mamou generate identical binaries,
with the exception of the postbyte in indexed addressing with
no offset: "ldb ,x" becomes "ldb 0,x" on mamou.

 make AS="mamou -r -q"

or

 make AS="lwasm -r --pragma=condundefzero,nodollarlocal"


Finally, switch to lwasm as default assembler.

Some of the binaries have been tested using xroar, for instance:

 xroar -machine coco -bas cocoroms/bas13.rom \
       -extbas cocoroms/extbas11.rom \
       -run hdbdos/hdbdw4cc2.wav