hdbdos: Replace "cond" with "IFDEF" or the equivalent "IFNE"
Because "cond" is not supported by lwasm. "IFDEF" is new in mamou.

Use the more common and intuitive "IFDEF" in the case of a single
symbol which is simply defined or not, and use "IFNE" on expressions.

Note that at one place the single symbol VREV is used but it is always
defined and should be evaluated as a number, hence IFNE is used here.

Also uppercase all "IFNE", "IFGT", "ELSE" and "ENDC".