hdbdos: Demystify the preamble and postamble for DECB files
The "source" file was just a list of FCBs. I have disassembled it and
commented it based on my interpretation. Note that I didn't write the
code in the first place, so YMMV.

The check for a $8cXX reset vector remains a mystery to me, I don't know
which ROMs that would be.