Open files in binary mode (for Windows)
On Windows the default text file mode might imply conversion
of end-of-line markers and interpretation of end-of-file characters.

This was in particular causing wrong detection of disk image type in
libcoco in some cases on Windows, with failing os9copy (error 248) as a
result.

Also fixed this in makewav and os9rdump.