os9format: Fix minimal bitmap size
This looks like a typo. After the check a few lines up, dd_tot must be 4
or higher, so it would never be zero here.

I believe the original thought was to make sure there is at least one
byte in the allocation bit map, in the case there are so few sectors
that there would be less than 8 bits in the allocation map, and the
divions would yield zero.

Since dd_map is never zero, bitmapSectors will also never be zero, so
we can remove the minimal bitmapSector check.