os9format: Really fix bitmap allocation size
Follow up to "Fix minimal bitmap size"

There was more than just a typo, but a thinko as well.
The byte length of the allocation bit map must always be
rounded up to the next byte, not only in the case of less
than one full byte.

In the calculation of clusters to allocate, we don't need
to round up the number of clusters because the number of
sectors has already been rounded up to a multiple of cluster
size.