libdecb: Fixup last HDB-DOS detection change
My recent change wrongly assumed the :num part is before the comma,
whereas it should be after the comma, and last in the path.

Also make it less strict so that it will allow offsets, e.g.
IMGFILE,PATH:23+0x58000

With this we are also sure there are no false DECB detection caused by
a windows drive designaton like "C:" in the image part before the comma.