libdecb/librbf: strncpy is not always the best solution
In particular, if we know the source string length and that there
is enough room in the destination, a simple memcpy is just as good
(when we don't need the extra zero padding of strncpy).
And compiler is happy...

In many cases our destination is a fixed length character array and
not a C string that would need terminating zeroes.

Also fixes a possible overflow issue in os9_rename() if the new name
was exactly 29 characters.