libdecb: Replace strndup() and fix paths using colon
The original code had q - p inverted by mistake, so the disk
image name was not cut off on a ":" delimiter.

strndup() is missing on MingW here, so replace it with
a series of standard functions.

Leave the corrected and more elegant strndup() version
in a comment for possible future use.