decb: Fix possible corruption of HDB-DOS virtual floppy images
These are always 35 tracks with 68 granules so don't look
for free granules beyond that.

Fixes an issue where images of some real floppies have their
granule table padded with $FF and we would count them as
free granules, making it possible to write above the floppy
image limit and overwrite other floppy images.

The meaning of the hdbdos_offset member was changed so that it
can flag a HDBDOS image with an offset of zero.

https://sourceforge.net/p/toolshed/bugs/39/