Fix HDBDOS sources to account for a change in LWASM.
LWASM in raw mode gets confused by RMB statements.
Sometimes it thinks it should generate 0 bytes into the binary.
That seems to be after it has already been generating code.

There's also a problem if you ORG to a differnt address
in the middle of generating raw bytes.   (It's OK to ORG
to the same address you are already at.)

First I tried making a change to LWASM, but it broke
something else.  It turns out I can fix the HDBDOS
sources instead, so the confusing situations do not occur.