Tim's refactor of track0 sectors for os9 format (#34)
* fix Track0 sector size in formatting Hard Drive.

It should not be different from the sector size on other tracks.

Also a unit test that breaks when the bug exists.

FIXES https://github.com/nitros9project/toolshed/issues/28

* Tim's refactor of track0 sectors for os9 format

This simplifies how the sectorsTrack0 variable is used, giving it an "unset" value of 0.

It moves a test that forks the "os9 format" command into a new file unittest/os9commandtest.c

Fixes the order of libraries, as needed in automated testing.

This PR was originally https://github.com/nitros9project/toolshed/pull/31 but the fundamental bug fix was made in https://github.com/nitros9project/toolshed/pull/29 so it has been adapted.